+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
-_No unreleased changes yet._
+- bug: update `auto-update-master.sh` to reset major tag when updating minor and patch versions
+- bug: refactor logic to ensure scheduled run defaults to dry-run = false, change switch to Boolean
+
 
 ## v1.0.1
 2026-02-10